refactor(tests): extract cst_symbol.rs's inline mod tests into a companion file
docs/agent-reference.md rule #10: tests live in companion *_tests.rs files,
not inline mod tests {} blocks -- this file already violated it before this
branch, and the last two commits added more tests to the same inline block
instead of catching it. Pure extraction (verbatim content, same #[test]
functions, same assertions) plus the standard 3-line stub; no behavior
change.
